Fix the auth flow
